Bugfix pullHic
After running .prepareInputs() which involves snapping ranges to bins, the adjusted ranges were not being used to calculate the expected matrix dimensions. This can sometimes cause a missmatch between the data used for enumerating bins, and the data that is extracted from the Hi-C file. Fixed by using the adjusted interactions to set the matrix dimensions.